+++ kivy/screens/tts.py
@@ -1,6 +1,7 @@
 # screens/tts.py
 from kivymd.uix.boxlayout import MDBoxLayout
 from kivymd.uix.label import MDLabel
+from kivymd.uix.spinner import MDSpinner
 from kivymd.uix.button import MDFillRoundFlatButton
 from kivymd.uix.textfield import MDTextField
 from kivymd.uix.scrollview import MDScrollView
@@ -9,13 +10,43 @@ from kivy.lang import Builder
 from kivy.properties import StringProperty, NumericProperty
 from kivy.metrics import dp, sp
 
+# get path details
+import sys
+import os
+if getattr(sys, 'frozen', False):
+    # Running as a PyInstaller bundle
+    base_path = sys._MEIPASS
+    noto_font = os.path.join(base_path, "data/fonts/NotoSans-Merged.ttf")
+else:
+    # Running in a normal Python environment
+    base_path = os.path.dirname(os.path.abspath(__file__))
+    noto_font = os.path.abspath(os.path.join(base_path, "..", "data/fonts/NotoSans-Merged.ttf"))
+
 Builder.load_string('''
 #:import parse_color kivy.parser.parse_color
+
+<TempSpinWait>:
+    id: temp_spin
+    orientation: 'horizontal'
+    adaptive_height: True
+    padding: dp(8)
+
+    MDLabel:
+        text: root.text
+        font_style: "Subtitle1"
+        adaptive_width: True
+        #theme_text_color: "Custom"
+        #text_color: "#f7f7f5"
+
+    MDSpinner:
+        size_hint: None, None
+        size: dp(14), dp(14)
+        active: True
 
 <UsrMsg>:
     mode: "fill"
     readonly: True
-    font_name: 'data/fonts/NotoSans-Merged.ttf'
+    font_name: root.noto_path
     multiline: True
     max_height: "200dp"
     size_hint_x: 0.7
@@ -46,7 +77,7 @@ Builder.load_string('''
 <MultiLingualTextField>:
     hint_text: "Type your text..."
     mode: "rectangle"
-    font_name: 'data/fonts/NotoSans-Merged.ttf'
+    font_name: root.noto_path
     multiline: True
     max_height: "200dp"
     input_type: 'text'
@@ -100,25 +131,36 @@ Builder.load_string('''
         MultiLingualTextField:
             id: chat_input
 
-        MDFillRoundFlatButton:
-            id: send_msg_button
-            text: "Send"
-            size_hint_x: 0.2
-            size_hint_y: 1
-            font_size: sp(18)
-            on_release: app.send_message(self, chat_input, chat_history_id)
+        MDIconButton:
+            icon: "send"
+            icon_size: sp(32)
+            pos_hint: {'center_y': 0.5}
+            theme_icon_color: "Custom"
+            icon_color: app.theme_cls.primary_color
+            on_release: app.send_message(self, chat_input)
 ''')
+
+class TempSpinWait(MDBoxLayout):
+    text = StringProperty("")
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
 
 class TtsResp(MDBoxLayout):
     id = StringProperty("")
 
 class MultiLingualTextField(MDTextField):
-    pass
+    noto_path = StringProperty()
+    def __init__(self, noto=noto_font, **kwargs):
+        super().__init__(**kwargs)
+        self.noto_path = noto
 
 class UsrMsg(MDTextField): # originally MDLabel
     id = StringProperty("")
     text = StringProperty("")
-    #font_style = "Subtitle1"
+    noto_path = StringProperty()
+    def __init__(self, noto=noto_font, **kwargs):
+        super().__init__(**kwargs)
+        self.noto_path = noto
 
 class TtsBox(MDBoxLayout):
     def __init__(self, **kwargs):